# swagger.yml
openapi: 3.0.0
info:
  title: Events Dapp
  description: An Event Management Dapp with Metamask Login Support
  version: 1.0.0
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid
  schemas:
    address:
      type: string
      maxLength: 42
      minLength: 42
      example: "0x9fB48802C9c9A187Df19AF823a792b909bec8576"
...
paths:
...
  /auth/nonce/{pubkey}:
    get:
      tags:
        - auth
      summary: Gets a nonce which should be signed with the User's private key to beat the challenge
      parameters:
        - name: pubkey
          in: path
          required: true
          description: public key/wallet address of the authenticating User
          schema:
            $ref: "#/components/schemas/address"
      responses:
        "200":
          description: successfully got the nonce for signing
  /auth/login:
    post:
      tags:
        - auth
      summary: Reads the provided public key with signed message and tries to log the User in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pubkey:
                  $ref: "#/components/schemas/address"
                signedmsg:
                  type: string
                  description: The nonce returned by the previous login step, signed by the user's private key
      responses:
        "200":
          description: "Successfully logged in"
          headers:
            Set-Cookie:
              $ref: "#/components/securitySchemes/cookieAuth"
        "401":
          description: "Failed to log in"